pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name            = "idosell-api"
version         = "0.8.1.1b1"
description     = "Python SDK for Idosell REST API"
readme          = "README.md"
license         = "MIT"
license-files   = ["LICENSE"]
requires-python = ">=3.10"
authors         = [
    {name = "Bartłomiej Moszyński", email = "bartekm@gmail.com" }
]
keywords        = ["idosell", "api", "ecommerce", "client", "sdk", "rest", "cms", "crm", "oms", "pim", "system", "wms"]
classifiers     = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Distributed Computing"
]

dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.1",
    "python-dateutil>=2.9.0",
    "python-dotenv>=1.1.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pylint>=3.3.8",
    "ruff>=0.13.1",
]

[project.urls]
Homepage        = "https://github.com/mozan/idosell-python"
Repository      = "https://github.com/mozan/idosell-python"
Issues          = "https://github.com/mozan/idosell-python/issues"
Documentation   = "https://github.com/mozan/idosell-python#readme"
